# [Routing] fix perf issue when dumping large number of routes